backend/requirements.txt
flask
openai-whisper
torch
moviepy
ffmpeg-python
opencv-python
numpy